Completing uint32_t change

+++ contracts/include/reward-series.hpp
@@ -46,7 +46,7 @@
         } else if (seriesName == "HEXAGONAL") {
             return series::HEXAGONAL;
         } else {
-            return std::vector<uint64_t>();
+            return std::vector<uint32_t>();
         }
             
     }//END getSeries()
